Remove ErrorCompiler::Input
